updating ollama script

+++ setup_ollama.sh
@@ -1,11 +1,20 @@
 #!/bin/bash
-export BACKEND=tgi
-wget -O - "https://raw.githubusercontent.com/zenpalio/vastai-pyworker/main/start_server.sh" | bash && (text-generation-launcher --model-id "$MODEL_ID" --json-output --port 5001 --hostname "0.0.0.0" &>> $MODEL_LOG)
 
-wget https://huggingface.co/Orenguteng/Llama-3.1-8B-Lexi-Uncensored-GGUF/resolve/main/Llama-3.1-8B-Lexi-Uncensored_F16.gguf 2>&1 | tee model_download.log
-# Define the template content
-echo "Creating the model file2"
-cat << 'EOF' > llama3-uncensored.modelfile
+# Download and run start_server.sh only if it doesn't exist
+if [ ! -f "start_server.sh" ]; then
+    wget -O start_server.sh "https://raw.githubusercontent.com/zenpalio/vastai-pyworker/main/start_server.sh"
+    bash start_server.sh && (text-generation-launcher --model-id "$MODEL_ID" --json-output --port 5001 --hostname "0.0.0.0" &>> $MODEL_LOG)
+fi
+
+# Download the model file only if it doesn't exist
+if [ ! -f "Llama-3.1-8B-Lexi-Uncensored_F16.gguf" ]; then
+    wget https://huggingface.co/Orenguteng/Llama-3.1-8B-Lexi-Uncensored-GGUF/resolve/main/Llama-3.1-8B-Lexi-Uncensored_F16.gguf 2>&1 | tee model_download.log
+fi
+
+# Create the model file only if it doesn't exist
+if [ ! -f "llama3-uncensored.modelfile" ]; then
+    echo "Creating the model file"
+    cat << 'EOF' > llama3-uncensored.modelfile
 FROM ./Llama-3.1-8B-Lexi-Uncensored_F16.gguf
 TEMPLATE """{{ if .System }}<|start_header_id|>system<|end_header_id|>
 
@@ -18,6 +27,7 @@ SYSTEM """You are Veronica a flirty girl of age 21"""
 PARAMETER num_ctx 8192
 PARAMETER num_gpu 99
 EOF
+fi
 
 # Create and run the model using ollama
 echo "Creating and running the model"
@@ -25,5 +35,3 @@ ollama serve &
 sleep 10
 ollama create llama3-uncensored -f llama3-uncensored.modelfile 
 ollama run llama3-uncensored
-
-# Export BACKEND variable and download/start the server script
